requirements.txt
llama-cpp-python
chromadb
sentence-transformers
PyMuPDF
streamlit
python-dotenv
